requirements.txt
Flask~=1.1
Pillow~=6.1